+++ flow-typed/Noteplan.js
@@ -352,6 +352,20 @@ type TDataStore = {
    * Get all regular, project notes.
    */
   +projectNotes: $ReadOnlyArray<TNote>,
+  /**
+  * Get all cached hashtags (#tag) that are used across notes.
+  * It returns hashtags without leading '#'.
+  * @type {[String]}
+  * Note: Available from NotePlan v3.6.0
+  */
+  +hashtags: $ReadOnlyArray < string >,
+  /**
+* Get all cached mentions (@name) that are used across notes.
+* It returns mentions without leading '@'.
+* Note: Available from NotePlan v3.6.0
+* @type {[String]}
+*/
+  +mentions: $ReadOnlyArray < string >,
 
   /**
    * Get or set settings for the current plugin (as a JavaScript object).
@@ -583,6 +597,10 @@ type PluginCommandObject = {
    * Name of the plugin this command belongs to (getter)
    */
   +pluginName: string,
+  /**
+   * List of optional argument descriptions for the specific command (getter). Use this if you want to invoke this command from another plugin to inform the user what he nees to enter for example.
+  */
+  +arguments: $ReadOnlyArray < string >
 }
 
 type PluginObject = {